+++ toolkit/Source/src/engine/Scene/ETHScene.h
@@ -73,7 +73,7 @@ public:
 	void ForceAllSFXStop();
 
 	void Update(
-		const unsigned long lastFrameElapsedTime,
+		const float lastFrameElapsedTime,
 		const ETHBackBufferTargetManagerPtr& backBuffer,
 		asIScriptFunction* onUpdateCallbackFunction);
